Send "geolocationServiceAvailable" with startup request
Instead of sending a separate request after the desktop has been created
on the UI, the information whether geolocation services are supported by
the browser is sent along with the startup request (in the
sessionStartupParams map). While the desktop is attached, this
information is available in PropertyMap.CURRENT.

The method attachGui() on the desktop's UI facade was changed. The deep
link target is no longer passed as an argument, because it is also
available in PropertyMap.CURRENT (consistency).

Change-Id: Ieb9f84c5f629687ae1013c02676d3bc168f50b43