Fix JS Error on ShowInvisibleColumnsForm resize
When the ShowInvisibleColumnsForm gets resized, JQuery UI will propagate
the event to the parent element and up until document.window! This
propagation can't be stopped in the resizable callback, because it is
called after the normal event handling (and regardless if the event
propagation got stopped). In order to stop the propagation we have to
use a normal resize event handler instead. (See last comment on
https://bugs.jqueryui.com/ticket/7514)

Note: Also changed the parent of ShowInvisibleColumnsForm to Desktop,
because otherwise the Form gets hidden by clicking on the Table.
180541

Change-Id: I158afe612d788e01835cd42da4265f13b29b6781
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/84094
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>