Menubar: Space between buttons is too large
The previous commit set the gap to 6px, to make the gaps between buttons
equal (gap between navigate buttons was 6px). We decided to change it to
10px because it looks better for regular buttons. In addition, the gap
between the navigate button block and the buttons on the right has been
increased to 20px to clearly separate these button block from other
buttons.

177767