Widget: setParent must not be allowed if parent is being removed
Usecase: Mobile client is used, search form is open. Search attribute is adjusted and search triggered by pressing search -> search form will close. The next time the same search form is opened, an "already rendered" exception occurs. 
Reason: Search form is still rendered because at the time the mobile popup is removed (which is after the animation), the form is not linked to the popup anymore. This is because the table was reloaded due to the search and the search form table control recreated which linked the search form to the table control again using setParent().