Fixed problem with touch smartfield popup.
The sovled problem occurred when a user typed 'Baden' into the textfield
within the smartfield popup. When the user clicked on a lookup row with
exactly the same text ('Baden'), the popup was closed, but the display
text in the field behind the popup did not change. That's because this
field never receives a displaytext changed event, because on the model
the displaytext was already 'Baden' at the time the row was clicked.
Thus the field did not trigger a property change event.