Fix layout of AbstractPageField
GroupBoxLayout.js added some hard coded "special" paddings (14/16px) in
certain cases (groupbox without border but with menu bar). This
implementation was wrong, as it added the paddings twice to the
preferred height after the first layout pass. This caused the UI to
"jump" if the form was relayouted after the initial layout (e.g.
changing the browser window size).

The special paddings seem to be unnecessary. The original issue would
have been better resolved by setting the correct model properties.
Therefore, the entire change (616b225899bf98c0d2fb44e99e11fed5a008855c)
was reverted.

Additionally, the AbstractPageField set a strange weightY of 0.001 on
the search form field. There is no reason (and no comment on) why this
should be necessary. It causes additional pixels to be added to the
search form's height, which are not necessary. This value was probably
introduced a long time ago to compensate for a now-gone bug in another
component. Therefore, the value was changed to 0 (= search form has
fixed height, table field uses the remaining space).

[CP from 6.1.x]

Change-Id: Ida2db7c614ee3fb6e4c654020b4661dac8004850