Don't use MultiSessionCookieStore in "bridge" mode
MultiSessionCookieStore is only required when the server tunnel uses
HTTP transport. When the org.eclipse.scout.rt.serverbridge module is
active, the tunnel invokes beans directly (because client and server are
running in the same VM). In this case, the
MultiSessionCookieStoreInstallListener must _not_ install the
MultiSessionCookieStore.

The MultiSessionCookieStoreInstallListener was moved from the *.ui.html
module to the *.client module, because the service tunnel is not part of
the UI, but part of the client model.

~179245

Change-Id: I9d1508493bca9cd5648a03432c15681bf8a243b8