Also consider parent-scopes when calculating sub-scopes
193632

Consider the following usecase:
Given a form containing 2 fields:
1. A normal button
2. A groupbox that inherits from a custom MyGroupBox template which
contains a custom field.
This custom field therefore exists in the primary type of MyGroupBox
(not in the form).

If an extension is created for each of the two fields a sub-scope is
created for the button (which has a higher level in the field tree).
Then this sub-scope is searched for further extensions when stepping
into the MyGroupBox template. But this scope only contains scope-items
of the form and no items of the MyGroupBox template. These would be in
the global scope which is untouched. 

Therefore when calculating the sub-scope for stepping into the
MyGroupBox it is not sufficient to only looking at the current sub-scope
(of the form) but we also have to consider the parent scopes because it
could be a template whose scope-items are only part of the global-scope.

Change-Id: I2efb05227f5a8e8dd69473e86bb112550231d56f
Reviewed-on: https://git.eclipse.org/r/85426
Tested-by: Hudson CI
Reviewed-by: Andi Bur <andi.bur@gmail.com>