Fix unload request in Firefox
When the window is unloaded, the 'unload' message should be sent to the
server, so the UI session can be disposed and detached from the desktop.
Because in Firefox, the "offline" variable is already set to true when
the unload handler is executed, a special condition must be added.
Otherwise, the UiSession will not be disposed.

173179

Change-Id: Ic75e8fa9fe03160129268f35fe2ac64a45a663e0