Removed ensureObjectType option from ObjectFactory
Just check if created object has an init function instead. This allows simpler usage of ObjectFactory with classes that do not have a init function.
Added objects#isFunction (delegates to jQuery#isFunction)
Added JasmineScout#createAdapterModel to prevent accidental modification of model object
ObjectFactory now always uses the objectType which is passed as objectType parameter
All this was required in order to set the default values for gridDataHints properly on GroupBox.js