App: Move object factory to prepare phase
Some objects are not extensible because they are created during parse
time (like scout.Device). At that time the object factory is not ready
and therefore replacing the device is not easily possible.

The app now has a prepare phase and triggers events on phase changes.
This gives the possibility to execute code before the session is
initialized, like creating the device using the object factory.

Also moved adapter specific code from the RemoteApp to the responsible
adapter (like TableAdapter)