Add AbstractConfiguredMomDelegate to reduce copy/paste code
The code in each MOM was mostly the same. To reuse this code from
multiple MOM classes, a new base class was added:
AbstractConfiguredMomDelegate.

Static isNullImplementor() method was moved from properties to MOM
classes (as isNullTransport()).

A single MomPlatformListener now destroys all MOM transports on platform
shutdown.

NullMomImplementor is now final, because it represents a non-extensible
sentinel value. By making it final, it can be checked using ==,
"instanceof" checks are not needed.

Change-Id: I8a91f562dceb3bde4e4f7ebff4aaa44711422e5a