Layout: Fix exception introduced with ced9d48
Exception occurred because validateLayout was called for proposalChooser
inside an invisible smart field popup. This happened because
layoutValidator was not used due to the call of validateLayout instead
of validateLayoutTree (smartfield popup is invisible in editable table
if not displayed in viewport)

Fixed by moving the invisible check to the HtmlComponent -> Makes the
layout validator optional again.

See ced9d48c8acbf09d41c32b90cf734ae3c1ca83f8