Revert 'ValidateLayout checks for invisible components (recursively)'
Tree and table layout may throw errors if htmlComp.validateLayout is
called directly without using layout validator. This is the case for
smart field popups and may be for other elements too.
The fix for this error is added for 6.1 branch. Since it is a major
change in the layout behavior which is too risky to be ported, we
decided to revert this commit.