Table: Add formatValue to column
A cell may have a value and a text (same as in java). This was used by
number columns, date columns, smart columns etc. String columns only
used the text. This saves a little amount of memory but is inconsistent
and confusing.

To create a custom column with a plain object as value, you would have
to copy & paste a lot of smart column code.

In order to fix these 2 issues, the code from smart column has been
refactored and moved to column. And every column has to set a value for
the cells.

Result: The column now has a function formatValue which creates the text based on the value (similar to ValueFields). It can either return the formatted value or a promise. This may be used by every column type, even custom columns.