Added JAX-WS Web service and port pooling strategy
Creating web service and port instances are expensive operations (at
least if the reference implementation Metro or the one bundled with the
JRE is used). Especially parsing the WSDL and XSD files as well as
building JAXB contexts is significant and is because of synchronization
even worse if they are performed in parallel.

This change adds a new default strategy for creating ports that is based
by a pool. Actually the pooled provider uses two pools, one for service
instances and another for port instances (which are created by a service
instance). A Scout transaction member keeps track of leased ports and
puts them back into the pool when the Scout transaction releases its
resources. Further, the transaction member ensures that the same port is
used within a transaction, once it has been leased.

Port instances are reset when they are put back to the pool. Some JAX-WS
implementations provide a suitable operation (i.e. Metro as well as the
RI bundled with Java 8). Otherwise the request context is cleansed as
good as possible. The corresponding JaxWsImplementorSpecifics can be
extended to customize the cleansing.

As a side effect of this change, the AbstractWebServiceClient does not
distinguish anymore between PortProducer and PortCache. Both are
IPortProvider strategies and the new PooledPortPorvider is just another
one, frankly the new default. Setting the configuration property
jaxws.consumer.portPool.enabled to false enables the previous behavior.

The internal state of the pools are reported on the diagnostics servlet.

Change-Id: Ie9f0bf33f6f55917dab4f9cf731607237b280918
Reviewed-on: https://git.eclipse.org/r/87999
Tested-by: Hudson CI
Reviewed-by: Andi Bur <andi.bur@gmail.com>