Improve JS only locale handling
LocalesLoader and TextsLoader create locales.json and texts.json to be
used for js only apps. The locales for these loaders may now be
configured using the new property scout.ui.locales.

The ui stores the locales in scout.locales. Typically the application
wants to provide a drop down box to switch the locale. This is now
possible by using locales.getAll.

On startup, the app tries to read the language of the navigator and
creates a locale. This is only possible, if scout.locales contains an
appropriate locale for that language. If this is not the case, a default
locale is used.