Fix option name ("withFocusContext" instead of "installFocusContext")
The popup option "installFocusContext" was stored manually to the
property "withFocusContext" in the init() method. Commit
a028c2b481a8d2fb0d66a9a9e0781ffab2b43e6f changed the property
initialization to the "new way", where a generic method is used to copy
all options to the current object. Because the option name did not match
the property name, the property had the wrong value. As a consequence,
the focus stayed on the date field when pressing TAB.

To fix this problem, the option was renamed to the same name as the
property.

Change-Id: I5ff4950bb42e5babb1545cba2e82a51bb89bcf8f