JsonAdapter: Always send owner in inspector mode
Until now the owner was only sent if it is the root adapter. This is
fine for the current UI, because it determines the owner automatically,
meaning it uses the parent (creator) as owner. When using another client
like JMEter the owner may be relevant, because there is no logic to
determine it automatically. Because sending the owner would break the UI
(the owner widget cannot be resolved at the time the model adapter is
created), the owner is only sent in inspector mode and ignored by the
UI.
Added also a property global to make it more explicit that only the
global state is essential and the owner only used for testing purpose.