Desktop: fix "openUri" in Chromium
Using a hidden <iframe> for OpenUri actions DOWNLOAD and OPEN does not
work in Chromium due to a CSP-related bug, see
https://bugs.chromium.org/p/chromium/issues/detail?id=663325

Instead, window.location.assign() can be used in Chromium. Other
browsers still use the <iframe> solution, because they cancel the ?poll
request when the current window's location is changed. This does not
seem to be the case in Chromium.

183131
203445

Change-Id: I8e74b9b76ea31eac21bba9e34042238ff6901332
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/88023
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>