Desktop: fix "openUri" in Chromium
Using a hidden <iframe> for OpenUri actions DOWNLOAD and OPEN does not
work in Chromium due to a CSP-related bug, see
https://bugs.chromium.org/p/chromium/issues/detail?id=663325

Instead, window.location.assign() can be used in Chromium. Other
browsers still use the <iframe> solution, because they cancel the ?poll
request when the current window's location is changed. This does not
seem to be the case in Chromium.

https://git.eclipse.org/r/#/c/88023/
183131
203445

Change-Id: I999bb7c6c64f0f60abe99c0cedec2cf2631fd6ff
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/88148
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>