Menu: Missing border on selected menu item
Use case: Menu item has toggleAction = true and selected = true. If the
menu containing this menu item is opened, the bottom border was not
colored properly because the class 'next-to-selected' was not set on the
next menu item.
Same behavior if the menu is is clicked but the menu stays open -> menu
item is selected but bottom border is missing.