Fix strange shift-click table selection after search
When a row is selected and the rows are replaced (e.g. by executing a
search), that row may still be selected. In this case, pressing
shift-click did not behave as expected, because the internal
"lastActionRow" marker was cleared. To fix is, assume the first or last
of the selected rows to be the lastActionRow.

176284

Change-Id: I35e8adc58ef6cacfea42a2cc04860fcc6c8e014c