Fix potential endless loop while setting focus (IE)
When the selection range is set on a string field, the field will always
receive a 'focus' event in IE. However, the field may be covert behind a
glasspane (i.e. if a modal dialog is open). In this case, the focus
manager would remove the focus from the field again, which somehow (only
on IE, it is not exactly clear why) causes an endless loop, i.e. the
string field will keep receiving 'focus' events.

Change-Id: I78645374a6557bd99d8dc200524e7eecdc8a95b1