fix errorMessage issue with detached tabItems
adding a tooltip (i.e. validation error message) to a field, contained
in a tabItem currently not displayed, caused the tooltip being displayed
at a invalid position.
When rendering a tooltip, the attached state of the anchor is now
checked. If not attached, the tooltip is stored in the detachHelper's
storage for later retrieval.

Change-Id: I6c0361cacca954ab63c176eb323ff551534a1076
Signed-off-by: Ralf Muri <ralf.muri@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/88564
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Tested-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>