Mobile: Fix iOS scrolling bug
IOS has problems with nested scrollable containers. Sometimes the outer
container goes completely white hiding the elements behind.
This happens with the following case: Main box is scrollable but there
are no scrollbars because content is smaller than container.
In the main box there is a tab box with a scrollable table. This table
has scrollbars.
If the width of the tab box is adjusted (which may happen if the tab
item is selected and eventually prefSize called), the main box will go
white.

This happens only if -webkit-overflow-scrolling is set to touch.
To workaround this bug the flag -webkit-overflow-scrolling will be
removed if the scrollable component won't display any scrollbars.