FocusContext: revert restoring scroll positions after focus
Restoring the previous scroll position after focusing an element breaks
automatic scrolling in editable tables or when jumping from the last
field in a form to the first.

The change is being reverted, because its original cause is not known
anymore. If it is found to be necessary at a later point in time, it can
still be restored from the Git history.

Change-Id: Ie86a79aa572bbe4fb339b037d76f440f66adb8e1