Value field: Update 'value' property in widget to fix async model/UI
When the displayText is updated form the server model, the 'value'
property on the UI must be updated as well. This is done in the  method
parseAndSetValue(). The method is called in the ValueFieldAdapter (~=
similar to the server's UIFacade, which also calls parseAndSetValue()).
The 'value' and 'displayText' properties may still be changed
individually in the widget. The acceptInput() method will use
parseAndSetValue() as well to set both properties.

This fixes the following problem:
1. Simple form with a string field and a AbstractResetButton
2. Enter 'abc' in the field
3. Press the reset button -> text from field is removed. This state
   _looks_ correct, but actually, the 'value' property has still the
   old value 'abc'.
4. Enter 'abc' again the field (same text as before).
5. Press the reset button again -> nothing happens (NOK). This is
   because the (old) _parseAndSetValue() method thinks that the value
   did not change and therefore does nothing.

Change-Id: Ie096268014753b5a780f0d8c187de7cd60d3d5f3