Fix for onRowMouseDown/Up in editable and sortable Table
When an editable Table is sorted by the UI Server all rows in the DOM
will be recreated after a cell-edit. Because of that, the comparison
between the MouseDownRow and MouseUpRow in _onRowMouseUp did an early
abort, which made it impossible to edit a cell and then click on a next
editable cell and begin editing with just one click.

173426

Change-Id: Ia7d050495df20c5c8c07a9ea4832ad6f7b3a51f6
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/89077
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>