Fix for onRowMouseDown/Up in editable and sortable Table
When an editable Table is sorted by the UI Server all rows in the DOM
will be recreated after a cell-edit. Because of that, the comparison
between the MouseDownRow and MouseUpRow in _onRowMouseUp did an early
abort, which made it impossible to edit a cell and then click on a next
editable cell and begin editing with just one click.

173426, https://git.eclipse.org/r/#/c/89077/

Change-Id: I2f9f06e329e08f05fac4b8ac48ceb17b2375fae8
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/89079
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>