Fixed bug with menu visibility in menu bar
The menu bar did not listen on property changes of menu items that were not rendered. Because the property change listener was added in the render function. When a menu was invisible from the start, and later the visibility of the menu changed to visible, the menu bar stayed invisible.