IDeviceTransformer: add transformPageTable() to transform the page table
Transforming the page table during transformPage() emits a warning by
AbstractPage, because the table should not be initialized during the
page's "execInitPage" phase. To resolve this issue, a new method
transformPageTable() was added that is called during the pages'
"execInitTable" phase.

To avoid confusion and to clearify that those methods may be called
multiple times, the methods transformPageDetailForm() and
transformPageDetailTable() were renamed to notifyPageDetailFormChanged()
and notifyPageDetailTableChanged().

Change-Id: Id12960a98625e74734c591de3f7128ae70baf432
Reviewed-on: https://git.eclipse.org/r/89339
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>