Cell: fix "editable" marker if not all cells in a column are editable
In the remote case, the default values are stripped from JSON before
sending it to the UI. Therefore, the model received for cells may miss
some properties, such as 'editable' if they are equal to the default
value. Because the column applies some of its own properties to the cell
if they are missing, the default values for type 'Cell' must be applied
again to the cell model before that.

Without this change, the cells of all rows would be marked as editable
if the column is editable, even if in the Java model, not all rows are
editable.

Change-Id: I5e931752632f000ec185b598ed7cb7b5a06e2587