Fix device.isTableAdditionalDivRequired() for browser zoom != 100%
If browser zoom is not set to 100% when the application loads,
device.isTableAdditionalDivRequired() may return true in modern
browsers, due to rounding errors. However, adding an additional DIV is
not necessary in those browsers and can influence the alignment of text.

To be on the safe side, a threshold of 5px is used.

[CP from 6.1.x]

Change-Id: Ia0a6d9242a533e0eaeb909575043c7877c9d623f