DefaultBeanInstanceProducer returned partially initialized beans
Old behavior / Problem:
The DefaultBeanInstanceProducer used to create application-scoped beans
concurrently. The bean which constructor invocation returned first was
the one used in the application, while other instances were discarded.
Methods annotated with @PostConstruct were invoked on the winner bean
only and to prevent potential deadlocks, the constructed but yet not
post-constructed bean was already returned to other requesting threads.
NPEs and unpredictable behavior was the consequence. Some of these
affected beans were rewritten, so that the @PostConstruct procedure was
migrated into an ensured-initialized style (i.e. every method that
required initialization of further resources first invoked a
synchronized and performed-only-once initialization method).

Change:
Only completely initialized beans are returned by the
DefaultBeanInstanceProducer or a BeanCreationException is thrown, if the
creation fails. If multiple threads are requesting a yet not initialized
bean concurrently, one thread is creating the bean and the other one
waits at most 90 seconds for the creation to complete. After exceeding
the timeout the two threads are considered blocking each other and a
BeanCreationException is thrown in the waiting thread.

188851

Change-Id: Ib998483a55fce5e949d6bbb6a9b6aaf196caddf1
Reviewed-on: https://git.eclipse.org/r/89498
Tested-by: Hudson CI
Reviewed-by: Hannes Mueller <hannes.mueller@bsi-software.com>