Session-Unload: navigator.sendBeacon() can't post "application/json"
The Content-Type "application/json" shall not be used with
navigator.sendBeacon() (see
https://www.mozilla.org/en-US/security/advisories/mfsa2015-03/ and
https://fetch.spec.whatwg.org/#cors-safelisted-request-header)

To fix this this a new UnloadRequestHandler is introduced. Clients that
support the Beacon API will send an empty POST-Request to
"/unload/[uiSessionId]". Clients that don't support the Beacon API will
still send a regular JSON-POST-Request to "/json?unload”
(JsonMessageRequestHandler).

Testing:
- This change can be tested in Chromium <55 by activating
"chrome://flags/#enable-experimental-web-platform-features" which will
throw a "Failed to execute 'sendBeacon' on 'Navigator': sendBeacon()
with a Blob whose type is not CORS-safelisted MIME type is disallowed
experimentally" when a JSON-POST-Request is sent to "/json?unload" via
navigator.sendBeacon().
- In Chromium >=55 this should be the default behavior.

204387

Change-Id: I3dcf4bdcf97b15ca0609c2fea03642c5885c9202
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/89422
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>