Use non-null theme names to improve use of cache keys
Observed behavior: Even if scout.ui.prebuild is set to true, the first
user request build the scripts and stylesheets again. This was because
the prebuilder uses the string "default" as theme name if no default
theme is configured. However, the first user request used null as theme
name in that case. This caused the cache keys to be not equal. The null
value was used to support loading of 'color.less' instead of
'color-default.less'.

This change ensures that the theme name is never null. "default" should
be used instead in the variables and caches. Only the ScriptFileBuilder
checks if the theme is equal to "default" and skips the theme-specific
loading part.

Change-Id: I5da9cc4bad852d4700f259956587690fe7202464