Write current theme to the model, if the model has no value
If the model (desktop) provides a theme, this theme should be used. If
for some reason, the desktop has no value set, but the session has a
current theme (e.g. from a cookie), this value should be written back to
the model. Otherwise, any code that reads the value from the model,
would not get the correct value (e.g. a simple options dialog).

Change-Id: I5ff41af2e6229babf28d1191650397c3ee4eb3fc