Tree.js: expand selected node in breadcrumb mode only
Observed behavior: when a tree node is selected but not expanded and the
browser window is refreshed, that node is suddenly rendered as expanded.
However, the node was only expanded in the UI, but not on the server
model (mismatch between model and UI)!

This expansion happened during _setDisplayStyle(), which was called in
the initial rendering. It was intended to be done _only_ in breadcrumb
mode. Instead, the if only checked for "truthyness" of the displayStyle
property. As a fix, the corresponding code was moved to the already
existing "if" branch. Additionally, the "else" branch was simplified.

Change-Id: I302fa68ded00c70f94c65aef30fbc9ebf87a914f