AbstractTree: collapseAll should collapse all sub nodes
collapseAll() should recursively collapse all nodes below the specified
node. If one of the nodes is already collapsed, the method must still
visit its children, because some of them may still be expanded.

Consider this example:

  A
  +-B
  | +-C
  | | +-D
  | +-E  
  +-F
    +-G
    
When the user manually clicks on the collapse handle of "B", the tree
looks like this:

  A
  +-B
  +-F
    +-G
    
However, the node "C" has still expanded=true, so the original state can
be restored if the node "B" is expanded again.

When collapseAll("A") is called, the algorithm previously stopped at
"B", because "B" is not expanded. It therefore missed collapsing "C",
which would be the expected behavior.

Unit test added.

Change-Id: I354148e17fc572409eb4259eb4f590c7abdb4849