TableTextUserFilter: preserve cursor position while typing
When typing text in the table text filter field (in the footer), the
cursor position should not be altered. Otherwise, it is not possible to
add or remove multiple characters at the beginning or in the middle of
an existing string.

To fix the issue, the automatic conversion to lower case is no longer
written back to the <input> element (text stays as it is entered).
Instead, the conversion to lower case is only applied in the
TableTextUserFilter.

203883

Change-Id: Ia12d762a261be1e750e6e408074182f08308e3a7
Reviewed-on: https://git.eclipse.org/r/89950
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>