ListBox/TreeBox: Update active/checked filter on value change
Active/checked filters may hide some nodes/rows from list and tree
boxes. In some cases, they are visible nevertheless, namely if they are
checked programmatically (setValue), e.g. while displaying data that was
registered at a previous time when those entries were visible.

Observed behavior: If an entry is inactive, but shown (because it is
part of the value) and the value is changed so something different, the
inactive (but know unchecked) entry stays visible in the list. Also if
the value is changed to include an inactive entry, the value is checked
but not made visible.

With this change, the filters are re-applied whenever the value is
synced to the tree (syncValueToTree).

The filters were already re-applied in syncTreeToValue(). The
implementation was changed to match the implementation in
syncValueToTree(). Instead of removing and re-adding the filters, they
are now just re-applied. Also this now happens inside the try{} block. 

JsonTable:
Removing a previously filtered row from the filter causes a "row
inserted" event in the UI (because filtered rows do not exist for the
UI). If the newly visible row is just inserted it will always be added
at the end of the table. However, in the model it might have any
position. To send the position of the new row to the UI, an artificial
"row order changed" event is generated in JsonTable.

169985

[CP from 6.1.x]

Change-Id: I12614154f2166a67974571874dc30683e9c21bec
Reviewed-on: https://git.eclipse.org/r/90932
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>