Fixed problem with smart-fields in editable tables.
We noticed that in an editable table the smart field did not had the
proper error status code (NOT_UNIQUE_ERROR_CODE). When setAndParseValue
has been called on the smart field only a generic ParsingFailedStatus
has been set. Because the smart field relies on specific error codes we
must also set the status error code on the VetoException and also on the
ParsingFailedStatus.