Set desktop reference on session early on
Because a lot of widgets want to register a listener on the desktop in their init phase, they access the desktop by calling 'this.session.desktop' that's why we need this instance as early as possible. When that happens they access a desktop which is not yet fully initialized. But anyway, it's already possible to attach a listener, for instance. Because of this line of code here, we don't have to set the variable in App.js, after the desktop has been created. Also note that Scout Java uses a different pattern to solve the same problem, there a VirtualDesktop is used during initialization. When initialization is done, all registered listeners on the virtual desktop are copied to the real desktop instance.