Fixed a JS error caused by the FocusManager/-Context
Without the bugfix a problem could occur when:
1. a form has been opened in a tab, the form has an editable table with a smart-column.
2. when the smart-column has been edited and the proposal chooser has been opened the user clicks on a page in the outline navigation tree
3. now the focusManager finds the focusContext for the table-cell-editor popup, but the $container of that popup/context has already been detached
--> causes a JS error

The bugfix removes the popup when the table is detached. This also removes the focusContext from the focus manager. We think that's a reasonable behavior because a user does not expect the cell editor to be still opened when he/she returns to the previously detached form/table.