TableControl: Always send "form = null" property change
When the form is closed, the table control is automatically unselected.
Because the table control's JsonProperty for the form is not sent to the
UI (because of the accept() filter), the UI keeps the reference to the
form in the table control. When the table control is selected again in
the UI, that (destroyed) form is re-rendered automatically, which causes
an error ("Widget is destroyed").

To fix the issue, the table control must send the form when either the
control is selected _or_ the form is null.

Change-Id: I45995c09e3def23ca294f901e406dd08d7facbca