Prevent possible null-access when opening ContextMenuPopup
When a new popup gets opened,  _onPopupOpen() will check for existing
popups and close them. Because in Tree.js/Table.js holds a reference to
the currently opened popup a null-access can happen if the
‘onClose’-handler gets fired which sets the reference to null. 
This change adds a check if a popup is already present. If so, it will
close the old popup before opening the new one.
204781

Change-Id: Ieeb2dd8fe42b6c60e2d21fa9bb24e1fce40a7c42
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/90411
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>