Prevent possible null-access when opening ContextMenuPopup
When a new popup gets opened,  _onPopupOpen() will check for existing
popups and close them. Because in Tree.js/Table.js holds a reference to
the currently opened popup a null-access can happen if the
‘onClose’-handler gets fired which sets the reference to null. 
This change adds a check if a popup is already present. If so, it will
close the old popup before opening the new one.
6.1: https://git.eclipse.org/r/#/c/90411/
204781

Change-Id: I39a3e58c008bf4ef92b1ae25d6868671ee6488df
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/90412
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>