Fixed bug with file upload on iOS in home screen mode
On iOS home screen mode nothing happens when user clicks on the "Choose file" button. The reason for this was that FileChooser.js triggers a click event on the INPUT type=file element. The fastclick library messed up that event. With the additional CSS class 'needsclick' we give fastclick a hint, that it should not mess with that specific click event. Fastclick seems to be confused because we're in the middle of a click event for the "Choose file" button, and the event handler triggers another click for an INPUT element (because if we'd click directly on the INPUT element, fastclick would find that it should better not do anything for that element).

See ticket #204703