ListBox/TreeBox: Update active/checked filter on value change
Active/checked filters may hide some nodes/rows from list and tree
boxes. In some cases, they are visible nevertheless, namely if they are
checked programmatically (setValue), e.g. while displaying data that was
registered at a previous time when those entries were visible.

Observed behavior: If an entry is inactive, but shown (because it is
part of the value) and the value is changed so something different, the
inactive (but know unchecked) entry stays visible in the list. Also if
the value is changed to include an inactive entry, the value is checked
but not made visible.

With this change, the filters are re-applied whenever the value is
synced to the tree (syncValueToTree).

The filters were already re-applied in syncTreeToValue(). The
implementation was changed to match the implementation in
syncValueToTree(). Instead of removing and re-adding the filters, they
are now just re-applied. Also this now happens inside the try{} block. 

169985

Change-Id: Ic7f8df1c77c708ff3525a0db8fd0e926f3f78e7a