JsonTable: Correctly position rows added via row filter
If a filtered row gets unfiltered by the model, a "row inserted" event
is generated for the UI. This causes the new row to be inserted at the
bottom of the table. In the server model however, the row's position
might be somewhere else.

This change generates an artificial "row order changed" event with the
new row order whenever a row gets removed from a table row filter.

JsonTree does not need that technique, because inserted nodes have the
"childNodeIndex" attribute, i.e. they know their position in the parent
node's child node list.

Change-Id: I1f0fe4f4c4ceb296f8bf224f2bb759df729a2062
Reviewed-on: https://git.eclipse.org/r/90663
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>