Table: Don't use disposed columns as context column
If a column is used as context column and will be disposed, it must not
be used as context column anymore. Otherwise the UI (JsonTable,
m_jsonColumns.get(value)) will throw a NPE because the context column is
not known anymore.

May be reproduced by reloading the page after calling
ITable#resetColumnConfiguration.