Events send to a not rendered proposalChooser cause exception
Prevent events to be delegated to a table smartfield proposalChooser
that is not rendered anymore. This did happen if a table contained a
smart column that is mandatory inside a form. If the form is saved
without the mandatory column filled a message box is shown. Clicking
inside this message box caused the proposalChooser to be disposed in the
UI but this was not send to the server. Events like pressing key down
were still delegated to the invisible proposalChooser which caused an
error. 
Additionally do not show proposalChooser on focusRequest automatically
on table columns as it is currently for other smartfields.

Change-Id: Iaabf6081c9b7421f2756dff680a2198b725d1695
Signed-off-by: Michael Rudolf <mru@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/91069
Tested-by: Hudson CI