JsonFormField: Unregister adapters with visibleGranted=false if possible
Observed behavior: A form inside a AbstractFormMenu with fields that are
set to visibleGranted=false during execInitField() caused UI errors when
property changes happened on that field: "Could not resolve event
targets"

Analysis has shown that the field adapters are created and sent to the
UI (although they should not), because during attachChildAdapters() the
form is not yet started, and therefore the execInitField() methods have
not been executed yet. When the parent group box's toJson() method is
called, another DisplayableFormFieldFilter is applied and this time, the
visibleGranted property has to correct value. Therefore, the adapters
are sent to the UI (and listeners are attached), but they are not be
connected to any other adapters. Therefore, they will not be created on
the UI and events that happen later will not find their target.

To fix the problem (not only in the AbstractFormMenu case, but also in
AbstractTableControl, and maybe at other places), the JsonFormField
automatically disposes itself when it finds out that "visibleGranted"
was set to false. (If visibleGranted was false in the beginning, the
adapter would not have been created and attached at all.) This only
works as long as the adapter was not sent to the UI. To get this
information, the current response now provides a "writable" flag and a
method to check if an adapter is contained in the response. After a
JsonResponse has been converted to JSON, it is no longer writable.

The same fix was applied to JsonAction.

Some unit tests were added.

204106

Change-Id: I19c61e09224f7a001f68867118ba6ccc74f4a3ae
Reviewed-on: https://git.eclipse.org/r/90847
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>