Fixed tree bug, with wrong insert order in visibleFlatList
This bug fix addresses two issues:

1. nodes have been inserted at the wrong index (always on top) when
nodes have been added to the tree in the + (lazy) mode.

2. nodes have been inserted at the wrong index (almost random) when
nodes have been expanded when tree was in + (lazy) mode.

The problem was solved by using the new function
_findInsertPositionInFlatList every time we call _addChildrenToFlatList
recursively. This was missing in the previous implementation and caused
"random" insert positions for inserted nodes in some situations. This
new function replaces the old _find function.