ValueField: set the value initially
If the value is provided initially, the display text has to be updated.

In the remote case the server mostly only sends a display text. So if a
display text is provided initially it has to be used and not overridden
by formatting the null value (this is the same logic as for table
cells).