Tooltip: Fix "jumping" on text change
When a tooltip gets rendered, it is done in two phases: first the
container is rendered at top:0/left:0. Second in postRender the actual
width and height of the tooltip get measured  and based on that
dimensions the tooltip gets positioned besides its $anchor-element.

When the text of a tooltip changes those two phases where executed
again. But the tooltip is not at the top-left corner again, which can
lead to a line break in the first phase. In the second phase the tooltip
gets repositioned, which can lead to the loss of the line break of phase
1. When this happens, the tooltip has a false height which means also
false positioning.

In order to fix that, the tooltip will be positioned at top:0/left:0
before re-rendering in those two phases.

203521

Change-Id: Ifa5f6d1b9de6a6ce0b79f61f13765a863da7a330
Signed-off-by: Michael Schmuki <michael.schmuki@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/91303
Tested-by: Hudson CI
Reviewed-by: Andre Wegmueller <awe@bsiag.com>