Mobile: Improve dialog visualization
Dialogs have a margin on each side, this is not necessary on small
screens. Additionally, it makes the second button stack so that only the
first button is visible, if there is another menu item.

Also removed resize handle on small screens. It does not look good, it
may not be used on touch devices, and it increases the dialog size which
has the effect that the body gets horizontally scrollable.

Also removed border on small screens if dialog is opened full screen.
Looks a little smoother.

177767